+++ lib/common/routespl.c
@@ -1119,7 +1119,7 @@ makeStraightEdges(graph_t *g, edge_t **edge_list, int e_cnt, int et,
 	    Ppoint_t pts[] = {dumber[0], dumber[1], dumber[2], dumber[3]};
 	    Ppolyline_t spl, line = {.pn = 4, .ps = pts};
 	    make_polyline (line, &spl);
-	    clip_and_install(e0, aghead(e0), spl.ps, spl.pn, sinfo);
+	    clip_and_install(e0, aghead(e0), spl.ps, (size_t)spl.pn, sinfo);
 	}
 	else
 	    clip_and_install(e0, aghead(e0), dumber, 4, sinfo);
